refactor: Content 컴포넌트 분리
